WebRTCInfo をハードコードする
