--- Sora/SDKInfo.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-/// :nodoc:
-public struct SDKInfo {
-    // Sora iOS SDK のバージョンを定義する
-    public static let version = "2021.1"
-}
